+++ apiphany-core/src/main/java/org/apiphany/security/AuthorizationHeaderProvider.java
@@ -14,5 +14,4 @@ public interface AuthorizationHeaderProvider {
 	 * @return the value for the Authorization header
 	 */
 	String getAuthorizationHeader();
-
 }